backend/requirements.txt
alembic==1.0.10
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
Django==2.2.1
django-cors-headers==3.0.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
idna==2.8
Mako==1.0.9
MarkupSafe==1.1.1
passlib==1.7.1
pycparser==2.19
pycryptodome==3.8.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
